feat: Add documentation for all public APIs
This commit adds comprehensive Dartdoc comments to all public classes, methods, and properties in the `lib/` directory. This improves the usability and maintainability of the library by providing clear explanations for all public-facing APIs.

The following changes were made:
- Added Dartdoc comments to all public members in `lib/src/*.dart`.
- Refactored several constructors to follow Dart best practices and improve code clarity.
- Fixed minor analysis issues and a test bug that were discovered during the documentation process.
- Updated the task file to mark the documentation requirement as complete.

+++ test/src/async_parallel_batch_flow_test.dart
@@ -6,10 +6,12 @@ void main() {
     test('should process a batch of inputs in parallel', () async {
       final flow = AsyncParallelBatchFlow<int, int>([
         AsyncNode(
-          (dynamic r) async => (r as Map<String, dynamic>)['input'] * 2,
+          (dynamic r) async =>
+              ((r as Map<String, dynamic>)['input'] as int) * 2,
         ),
         AsyncNode(
-          (dynamic r) async => (r as Map<String, dynamic>)['input'] * 3,
+          (dynamic r) async =>
+              ((r as Map<String, dynamic>)['input'] as int) * 3,
         ),
       ]);
